# api/openapi.yaml
openapi: 3.1.0
info:
  title: Go Chi Boilerplate API
  version: 1.0.0
  description: Public contract for the boilerplate example API.
servers:
  - url: http://localhost:8080
tags:
  - name: Health
  - name: Widgets
paths:
  /health/live:
    get:
      operationId: getLiveness
      summary: Check process liveness
      description: Returns success when the API process can serve HTTP requests.
      tags: [Health]
      responses:
        "200":
          description: The process is alive.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
  /health/ready:
    get:
      operationId: getReadiness
      summary: Check service readiness
      description: Returns success when every required dependency is available.
      tags: [Health]
      responses:
        "200":
          description: Required dependencies are available.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"
              example:
                status: ok
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /v1/widgets:
    get:
      operationId: listWidgets
      summary: List widgets
      description: Returns a bounded cursor-based page of widgets.
      tags: [Widgets]
      parameters:
        - name: limit
          in: query
          description: Maximum number of widgets to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 20
        - name: cursor
          in: query
          description: Opaque continuation cursor returned by the previous page.
          schema:
            type: string
            minLength: 1
            pattern: "^[A-Za-z0-9_-]+$"
      responses:
        "200":
          description: A page of widgets.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WidgetList"
              example:
                items:
                  - id: 1
                    name: first widget
                    createdAt: "2026-07-10T10:00:00Z"
                    updatedAt: "2026-07-10T10:00:00Z"
                limit: 20
                nextCursor: MjAyNi0wNy0xMFQxMDowMDowMFp8MQ
        "400":
          $ref: "#/components/responses/BadRequest"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
    post:
      operationId: createWidget
      summary: Create a widget
      description: Validates and creates one widget.
      tags: [Widgets]
      requestBody:
        required: true
        description: Widget attributes to create.
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 120
                  example: first widget
                  x-oapi-codegen-extra-tags:
                    validate: required,max=120
            example:
              name: first widget
      responses:
        "201":
          description: Widget created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
              example:
                id: 1
                name: first widget
                createdAt: "2026-07-10T10:00:00Z"
                updatedAt: "2026-07-10T10:00:00Z"
        "400":
          $ref: "#/components/responses/BadRequest"
        "413":
          $ref: "#/components/responses/RequestTooLarge"
        "415":
          $ref: "#/components/responses/UnsupportedMediaType"
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
  /v1/widgets/{widgetID}:
    parameters:
      - $ref: "#/components/parameters/WidgetID"
    get:
      operationId: getWidget
      summary: Get a widget
      description: Returns one widget by its identifier.
      tags: [Widgets]
      responses:
        "200":
          description: Widget found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Widget"
              example:
                id: 1
                name: first widget
                createdAt: "2026-07-10T10:00:00Z"
                updatedAt: "2026-07-10T10:00:00Z"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
    delete:
      operationId: deleteWidget
      summary: Delete a widget
      description: Permanently removes one widget by its identifier.
      tags: [Widgets]
      responses:
        "204":
          description: Widget deleted.
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
        "504":
          $ref: "#/components/responses/GatewayTimeout"
components:
  parameters:
    WidgetID:
      description: Unique widget identifier.
      name: widgetID
      in: path
      required: true
      schema:
        type: integer
        format: int64
        minimum: 1
  schemas:
    HealthResponse:
      description: Health status returned by probe endpoints.
      type: object
      additionalProperties: false
      required: [status]
      properties:
        status:
          type: string
          const: ok
          example: ok
    Widget:
      description: Public widget representation.
      type: object
      additionalProperties: false
      required: [id, name, createdAt, updatedAt]
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: first widget
        createdAt:
          type: string
          format: date-time
          example: "2026-07-10T10:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-07-10T10:00:00Z"
    WidgetList:
      description: Bounded page of widgets.
      type: object
      additionalProperties: false
      required: [items, limit, nextCursor]
      properties:
        items:
          description: Widgets in the current page.
          type: array
          items:
            $ref: "#/components/schemas/Widget"
          example:
            - id: 1
              name: first widget
              createdAt: "2026-07-10T10:00:00Z"
              updatedAt: "2026-07-10T10:00:00Z"
        limit:
          description: Page size applied to this response.
          type: integer
          format: int32
          minimum: 1
          maximum: 100
          example: 20
        nextCursor:
          description: Opaque cursor for the next page, or null on the final page.
          type: [string, "null"]
          minLength: 1
          pattern: "^[A-Za-z0-9_-]+$"
          example: MjAyNi0wNy0xMFQxMDowMDowMFp8MQ
    ErrorResponse:
      description: Stable client-safe API error envelope.
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          description: Client-safe error information.
          type: object
          additionalProperties: false
          required: [code, message]
          properties:
            code:
              type: string
              example: invalid_request
            message:
              type: string
              example: The request is invalid.
            details:
              type: object
              additionalProperties: true
  responses:
    BadRequest:
      description: The request is invalid.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: invalid_request
              message: The request is invalid.
    UnprocessableEntity:
      description: The request violates input or business constraints.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: validation_failed
              message: Request validation failed.
              details:
                fields:
                  - field: name
                    rule: required
    RequestTooLarge:
      description: The request body exceeds the configured size limit.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: request_too_large
              message: The request body is too large.
    UnsupportedMediaType:
      description: The request body is not JSON.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: unsupported_media_type
              message: Content-Type must be application/json.
    NotFound:
      description: The resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: widget_not_found
              message: Widget was not found.
    ServiceUnavailable:
      description: A required dependency is unavailable.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: not_ready
              message: Service is not ready.
    GatewayTimeout:
      description: The request exceeded the configured request deadline.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: request_timeout
              message: The request took too long to complete.
    InternalServerError:
      description: An unexpected internal error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            error:
              code: internal_error
              message: An internal error occurred.