# internal/httpapi/apigen/types.gen.go
// Package apigen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package apigen

import (
	"time"
)

// Defines values for HealthResponseStatus.
const (
	Ok HealthResponseStatus = "ok"
)

// Valid indicates whether the value is a known member of the HealthResponseStatus enum.
func (e HealthResponseStatus) Valid() bool {
	switch e {
	case Ok:
		return true
	default:
		return false
	}
}

// ErrorResponse Stable client-safe API error envelope.
type ErrorResponse struct {
	// Error Client-safe error information.
	Error struct {
		Code    string                  `json:"code"`
		Details *map[string]interface{} `json:"details,omitempty"`
		Message string                  `json:"message"`
	} `json:"error"`
}

// HealthResponse Health status returned by probe endpoints.
type HealthResponse struct {
	Status HealthResponseStatus `json:"status"`
}

// HealthResponseStatus defines model for HealthResponse.Status.
type HealthResponseStatus string

// Widget Public widget representation.
type Widget struct {
	CreatedAt time.Time `json:"createdAt"`
	ID        int64     `json:"id"`
	Name      string    `json:"name"`
	UpdatedAt time.Time `json:"updatedAt"`
}

// WidgetList Bounded page of widgets.
type WidgetList struct {
	// Items Widgets in the current page.
	Items []Widget `json:"items"`

	// Limit Page size applied to this response.
	Limit int32 `json:"limit"`

	// NextCursor Opaque cursor for the next page, or null on the final page.
	NextCursor *string `json:"nextCursor"`
}

// WidgetID defines model for WidgetID.
type WidgetID = int64

// BadRequest Stable client-safe API error envelope.
type BadRequest = ErrorResponse

// GatewayTimeout Stable client-safe API error envelope.
type GatewayTimeout = ErrorResponse

// InternalServerError Stable client-safe API error envelope.
type InternalServerError = ErrorResponse

// NotFound Stable client-safe API error envelope.
type NotFound = ErrorResponse

// RequestTooLarge Stable client-safe API error envelope.
type RequestTooLarge = ErrorResponse

// ServiceUnavailable Stable client-safe API error envelope.
type ServiceUnavailable = ErrorResponse

// UnprocessableEntity Stable client-safe API error envelope.
type UnprocessableEntity = ErrorResponse

// UnsupportedMediaType Stable client-safe API error envelope.
type UnsupportedMediaType = ErrorResponse

// ListWidgetsParams defines parameters for ListWidgets.
type ListWidgetsParams struct {
	// Limit Maximum number of widgets to return.
	Limit *int32 `form:"limit,omitempty" json:"limit,omitempty"`

	// Cursor Opaque continuation cursor returned by the previous page.
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`
}

// CreateWidgetJSONBody defines parameters for CreateWidget.
type CreateWidgetJSONBody struct {
	Name string `json:"name" validate:"required,max=120"`
}

// CreateWidgetJSONRequestBody defines body for CreateWidget for application/json ContentType.
type CreateWidgetJSONRequestBody CreateWidgetJSONBody